requirements.txt
django
django-environ
django-filter
djangorestframework
djangorestframework-jsonapi
mysqlclient
sqlalchemy
